Handle Angular routes with a listener
it seems more idiomatic than being hardcoded in the Kernel.
Also it's probably a good idea to not bypass all listener
(eg: what if we add some security or monitoring listeners?)
Also it benefits from the ResponseListener behavior.

Using twig for this is probably overkill. It's a workaround
to be able to access the location of the static file out of the box
but it would probably be possible to pass it to the listener
through a config file